% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogetest.R
\name{homogetest}
\alias{homogetest}
\title{homogetest}
\usage{
homogetest(resp_var, factores, df)
}
\arguments{
\item{resp_var}{is the name of the response variable}

\item{factores}{a vector with the names of the variables used as factor}

\item{df}{A \code{dataframe} object with a response variable and one or more
independent variables}
}
\value{
A \code{data frame} with several variables:
\itemize{
\item \code{fk_stat} value of the Fligner-Killeen statistic
\item \code{fk_pvalue} the p-value for the test
\item \code{lev_stat} value of the Levene statistic
\item \code{lev_pvalue} the p-value for the test}
}
\description{
Check homogeneity of variances of different groups using several tests
}
\details{
The function computes the Levene's test for homogeneity of variance across
groups (\code{\link[car]{leveneTest}}) and Fligner-Killeen Test of
Homogeneity of Variances (\code{\link[stats]{fligner.test}} for each of
the variables included in \code{factores} vector.
}
\examples{
\dontrun{
factores <- c('disturb_year', 'site', 'interaction(disturb_year, site)')
homogetest(resp_var = 'rs', factores = factores, df = evires)
}
}
