% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapirosNormal.R
\name{shapirosNormal}
\alias{shapirosNormal}
\title{shapirosNormal}
\usage{
shapirosNormal(df, resp_var, factor_vars)
}
\arguments{
\item{df}{A \code{dataframe} object with a response variable and one or more
independent variables}

\item{resp_var}{is the name of the response variable}

\item{factor_vars}{a vector with the names of the variables used as factor
variables}
}
\description{
Compute the normality of a subset of data. The function split the data into
groups (defined by \code{factor_vars}) and compute the normality of each
group using a shapiro test
}
