% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapirosNormal.R
\name{shapirosNormal}
\alias{shapirosNormal}
\title{shapirosNormal}
\usage{
shapirosNormal(df, resp_var, factor_vars)
}
\arguments{
\item{df}{A \code{dataframe} object with a response variable and one or more
independent variables}

\item{resp_var}{is the name of the response variable}

\item{factor_vars}{a vector with the names of the variables used as factor
variables}
}
\value{
A \code{data frame} with several variables:
\itemize{
\item \code{statistic} is the value of the Shapiro-Wilk statistic
\item \code{p_value} an approximate p-value for the test. See
\code{\link[stats]{shapiro.test}}}
}
\description{
Compute the normality of a subset of data. The function virtually splits the
data into groups (defined by \code{factor_vars}) and compute the normality of
 each group using a shapiro test (see \code{\link[stats]{shapiro.test}})
}
